fix(hub): Remove circular dep between hub and interface
```
$ madge --circular src/index.ts
Processed 5 files (1s) (2 warnings)

✖ Found 2 circular dependencies!

1) hub.ts > interfaces.ts
```

To remove this, the interface.ts was removed and all of it's
types were moved into hub.ts
