+++ packages/hub/src/hub.ts
@@ -23,7 +23,6 @@ import {
 } from '@sentry/types';
 import { consoleSandbox, dateTimestampInSeconds, getGlobalObject, isNodeEnv, logger, uuid4 } from '@sentry/utils';
 
-import { Carrier, DomainAsCarrier, Layer } from './interfaces';
 import { Scope } from './scope';
 import { Session } from './session';
 
@@ -48,6 +47,47 @@ const DEFAULT_BREADCRUMBS = 100;
  * `maxBreadcrumbs` option cannot be higher than this value.
  */
 const MAX_BREADCRUMBS = 100;
+
+/**
+ * A layer in the process stack.
+ * @hidden
+ */
+export interface Layer {
+  client?: Client;
+  scope?: Scope;
+}
+
+/**
+ * An object that contains a hub and maintains a scope stack.
+ * @hidden
+ */
+export interface Carrier {
+  __SENTRY__?: {
+    hub?: Hub;
+    /**
+     * Extra Hub properties injected by various SDKs
+     */
+    integrations?: Integration[];
+    extensions?: {
+      /** Hack to prevent bundlers from breaking our usage of the domain package in the cross-platform Hub package */
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      domain?: { [key: string]: any };
+    } & {
+      /** Extension methods for the hub, which are bound to the current Hub instance */
+      // eslint-disable-next-line @typescript-eslint/ban-types
+      [key: string]: Function;
+    };
+  };
+}
+
+/**
+ * @hidden
+ * @deprecated Can be removed once `Hub.getActiveDomain` is removed.
+ */
+export interface DomainAsCarrier extends Carrier {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  members: { [key: string]: any }[];
+}
 
 /**
  * @inheritDoc
